R/C input fixes, altimeter work
Added on-board pressure to altitude conversion in the Sensors object.
Fixed signs and names of R/C inputs to be more standardized.  Added code
to Config app to generate pressure to altitude lookup table.  Config app
now shows Altitude in meters from the on-board controller calc.  Added a
graph to the window so I could see variations.