removed patch rev number
I removed the patch rev number because we keep forgetting to update it, and it's not really necessary since the version number is displayed by groundstation. While It is nice to have the rev number (or commit hash) in the file, it's not worth it if it's going to be frequently incorrect. I also removed the list of "in development" since that's well encompassed in the Github issues and active dev branch now.